+++ src/tags/mapstructure.ts
@@ -2,15 +2,14 @@ import * as vscode from 'vscode';
 import { LineStruct } from '../completion';
 import { generateCompletionItem, generateFlagCompletionItems, CompletionItems } from './util';
 
-const formItems = [
-    (name: string) => `form:"${name}"`,
-    (name: string) => `form:"${name},omitempty"`,
+const mapstructureItems = [
+    (name: string) => `mapstructure:"${name}"`,
 ];
 
 const tagDelimiter: string = ",";
-const tagFlags: string[] = ["omitempty"];
+const tagFlags: string[] = ["omitempty", "remain", "squash"];
 
-export function generateFormCompletion(names: string[], ls: LineStruct): vscode.CompletionItem[] {
+export function generateMapstructureCompletion(names: string[], ls: LineStruct): vscode.CompletionItem[] {
     let items = new CompletionItems;
 
     if (ls.exactTagType) {
@@ -21,12 +20,10 @@ export function generateFormCompletion(names: string[], ls: LineStruct): vscode.
     }
 
     for (let name of names) {
-        for (let f of formItems) {
+        for (let f of mapstructureItems) {
             items.push(generateCompletionItem(f(name), ls));
         }
     }
-
-    items.push(generateCompletionItem('form:"-"', ls));
 
     return items.items;
 }